+++ packages/stylelint-config/rules/scss.js
@@ -3,64 +3,50 @@ module.exports = {
   rules: {
     /**
      * @-else
-    */
-
+     */
     /**
      * @-extend
-    */
-
+     */
     /**
      * @-function
-    */
-
+     */
     /**
      * @-if
-    */
-
+     */
     /**
      * @-import
-    */
-
+     */
     /**
      * @-mixin
-    */
-
+     */
     /**
      * @-rule
-    */
-
+     */
     /**
      * $-variable
-    */
-
+     */
     /**
      * %-placeholder
-    */
-
+     */
     /**
      * //-comment
      */
-
     /**
      * Declaration
-    */
-
+     */
     /**
      * Media feature
-    */
-
+     */
     /**
      * Operator
-    */
-
+     */
     /**
      * Partial
-    */
-
+     */
     /**
      * Selector
-    */
+     */
     // Disallow redundant nesting selectors (&).
     // 'scss/selector-no-redundant-nesting-selector': true
   }
-}
+};