+++ packages/tslint-config-angular/rules/angular.js
@@ -1,4 +1,0 @@
-module.exports = {
-  extends: 'tslint-angular',
-  rules: {}
-}