+++ packages/tslint-config-react/rules/a11y.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: 'tslint-microsoft-contrib',
-  rules: {
-    'react-a11y-img-has-alt': true
-  }
-}