+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog for @omnious/eslint-config-react
 
-## 0.0.1 (Apr 10, 2018)
+## 0.0.2-rc0 (Apr 11, 2018)
+
+* Dependency update `@omnious/eslint-config@0.0.2-rc0`
+* Complete `eslint-plugin-react`
+* Complete `eslint-plugin-jsx-a11y`